requirements.txt
Flask==3.0.0
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
psycopg-binary==2.9.9
psycopg2==2.9.9
cryptography==41.0.7